changed third

+++ Tre.Modules/Tre.Libraries/Tre.FileSystem/DeleteFile/delete.ts
@@ -1,5 +1,8 @@
 "use strict";
 import fs from 'node:fs';
-export default function (dir: string = ''): void {
-    return fs.unlinkSync(dir);
-};
+export default function (dir: string): void {
+    if(fs.existsSync(dir)){
+        fs.unlinkSync(dir);    
+    };
+    return;
+}
